tried moving all files to different path
