+++ fileIO.js
@@ -3,12 +3,12 @@ var fs = require('fs');
 const identity = (x) => x;
 
 exports.write_count_to_file = function(data) {
-  fs.writeFileSync(`${__dirname}/../db/counts.json`, JSON.stringify(data));
+  fs.writeFileSync('./db/counts.json', JSON.stringify(data));
 }
 
 exports.write_to_file = function(id, session, callback = identity) {
   var obj;
-  var data = fs.readFileSync(`${__dirname}/../db/data.json`, 'utf8');
+  var data = fs.readFileSync('./db/data.json', 'utf8');
   obj = JSON.parse(data);
   if(obj.hasOwnProperty(id)) {
     let lst = obj[id];
@@ -28,7 +28,7 @@ exports.write_to_file = function(id, session, callback = identity) {
     obj[id] = [];
     obj[id].push(session);
   }
-  fs.writeFileSync(`${__dirname}/../db/data.json`, JSON.stringify(obj));
+  fs.writeFileSync('./db/data.json', JSON.stringify(obj));
 
   if(typeof callback === 'function') {
     callback(obj);
